Updated app to use ghzh-clients library
* Deleted GitHubClient code
* Deleted ZenHubClient code
* Moved the GitHubClient and ZenHubClient to the ghzh-clients library.
* Moved client code located in `__init__.py` to `base.py` files.
